chore: reorganizing and creating dev container
