+++ .devcontainer/devcontainer.json
@@ -1,0 +1,16 @@
+{
+  "dockerComposeFile": "docker-compose.yaml",
+  "service": "dev-app",
+  "workspaceFolder": "/workspace",
+  "postCreateCommand": "bin/workspace-init",
+  "settings": {
+    "debug.javascript.usePreview": false
+  },
+  "extensions": [
+    "dbaeumer.vscode-eslint",
+    "eamodio.gitlens",
+    "github.copilot",
+    "esbenp.prettier-vscode",
+    "zxh404.vscode-proto3"
+  ]
+}
