+++ .devcontainer/docker-compose.yaml
@@ -1,0 +1,28 @@
+version: '3.7'
+
+volumes:
+  postgres:
+
+services:
+  dev-app:
+    image: mcr.microsoft.com/vscode/devcontainers/typescript-node:14
+    command: tail -f /dev/null
+    user: 1000:1000
+    network_mode: service:db-postgres
+    env_file:
+      - ../.env.development
+      - ../.env
+    volumes:
+      - ..:/workspace:cached
+
+  db-postgres:
+    image: postgres
+    container_name: analytics-db-postgres
+    env_file:
+      - ../.env.development
+      - ../.env
+    restart: unless-stopped
+    volumes:
+      - type: volume
+        source: postgres
+        target: ${POSTGRES_DATA:-/data/postgres}
