+++ .env
@@ -1,3 +1,20 @@
+# SECRETS
+# --------------------------------------------------------------------------------------------------
+# GITHUB_TOKEN - A personal access token for GitHub. It is used to install our private packages. REQUIRED
+
 # GRPC
 # --------------------------------------------------------------------------------------------------
 GRPC_PACKAGES='analytics'
+# GRPC_PORT - The port to run the GRPC server on. Default value: 9090
+
+# LOGGING
+# --------------------------------------------------------------------------------------------------
+# LOGGING_LOGLEVEL - The level of logging you want to use while launching the applcation. Default value: ERROR
+# LOGGING_SERVICE_NAME - The name of the service used in our logger. Default value: analytics@unknown
+
+# POSTGRES
+# --------------------------------------------------------------------------------------------------
+# POSTGRES_USER - The user in our local Postgres database. REQUIRED WHILE IN DEVCONTAINER
+# POSTGRES_PASSWORD - The password used in our local Postgres database. REQUIRED WHILE IN DEVCONTAINER
+# POSTGRES_DB - The database name in our local Postgres database. REQUIRED WHILE IN DEVCONTAINER
+# POSTGRES_DATA - The path to the Postgres data directory. OPTIONAL
