+++ .env.development
@@ -1,3 +1,9 @@
 # LOGGING
 # --------------------------------------------------------------------------------------------------
 LOGGING_LOGLEVEL='DEBUG'
+
+# POSTGRES
+# --------------------------------------------------------------------------------------------------
+POSTGRES_USER='analytics'
+POSTGRES_PASSWORD='changeme'
+POSTGRES_DB='analytics'
