--- .env.example
@@ -1,9 +1,0 @@
-# GRPC
-# --------------------------------------------------------------------------------------------------
-# GRPC_PORT - The port where our gRPC server will launch - Default: 9090
-# GRPC_PACKAGES - A comma-separated list of all gRPC applications you want to launch - REQUIRED
-#
-# LOGGING
-# --------------------------------------------------------------------------------------------------
-# LOGGING_LOGLEVEL - The log level for the application - Default: ERROR, Allowed values: DEBUG, INFO, WARN, ERROR
-# LOGGING_SERVICE_NAME - The name of the service that will be used for the logger - Default: "analytics@unknown"
