+++ src/app.module.ts
@@ -1,8 +1,8 @@
 import { Module } from '@nestjs/common'
 
-import { GRPCConfigModule } from './config/grpc/grpc.module'
-import { LoggingConfigModule } from './config/logging/logging.module'
-import { GRPCModule } from './grpc/grpc.module'
+import { GRPCConfigModule } from './infrastructure/config/grpc/grpc.module'
+import { LoggingConfigModule } from './infrastructure/config/logging/logging.module'
+import { GRPCModule } from './infrastructure/grpc/grpc.module'
 
 @Module({
   imports: [GRPCModule, GRPCConfigModule, LoggingConfigModule],
