+++ src/infrastructure/config/logging/logging.environment-schema.ts
