+++ src/infrastructure/grpc/analytics.proto
