+++ src/main.ts
@@ -4,10 +4,10 @@ import { MicroserviceOptions } from '@nestjs/microservices'
 import { buildLogger } from '@lib/logger'
 
 import { AppModule } from './app.module'
-import { GRPCConfigModule } from './config/grpc/grpc.module'
-import { GRPCConfigProvider } from './config/grpc/grpc.provider'
-import { LoggingConfigModule } from './config/logging/logging.module'
-import { LoggingConfigProvider } from './config/logging/logging.provider'
+import { GRPCConfigModule } from './infrastructure/config/grpc/grpc.module'
+import { GRPCConfigProvider } from './infrastructure/config/grpc/grpc.provider'
+import { LoggingConfigModule } from './infrastructure/config/logging/logging.module'
+import { LoggingConfigProvider } from './infrastructure/config/logging/logging.provider'
 
 async function bootstrap() {
   const grpcConfigContext = await NestFactory.createApplicationContext(GRPCConfigModule)
@@ -24,8 +24,8 @@ async function bootstrap() {
   )
 
   grpc.useLogger(logger)
-
   await grpc.listen()
+
   logger.log(`GRPC server listening on port ${grpcConfig.port}`)
 }
